add campground permissions

+++ routes/campgrounds.js
@@ -66,10 +66,16 @@ router.get(
   '/:id/edit',
   isLoggedIn,
   catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id);
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
     if (!campground) {
       req.flash('error', 'Cannot found a campground...');
       return res.redirect('/campgrounds');
+    }
+
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You do not have permission to do that!!');
+      return res.redirect(`/campgrounds/${id}`);
     }
     res.render('campgrounds/edit', { campground });
   })
@@ -82,7 +88,12 @@ router.put(
   validateCampground,
   catchAsync(async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, {
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You do not have permission to do that!!');
+      return res.redirect(`/campgrounds/${id}`);
+    }
+    const camp = await Campground.findByIdAndUpdate(id, {
       ...req.body.campground,
     });
     req.flash('success', 'Successfully updated a campground!!');
@@ -96,6 +107,11 @@ router.delete(
   isLoggedIn,
   catchAsync(async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You do not have permission to do that!!');
+      return res.redirect(`/campgrounds/${id}`);
+    }
     await Campground.findByIdAndDelete(id);
     req.flash('success', 'Successfully deleted a campground...');
     res.redirect('/campgrounds');
